Principal Components of a "bigSNP".

Get k or all Principal Components (PCs) of a bigSNP

PCA.bigSNP(x, block.size, k = NULL, ind.train = seq(nrow(X)),
  scaling = "none", thr.eigval = 0.001, use.Eigen = TRUE,
  progress = TRUE)

Arguments

x

A bigSNP.

k

Number of PCs to compute. Default is all.

scaling

Scaling to be used, either one of

"none",

"center": center columns,

"scale": center columns and divide by sd,

"binom": use \((x-2p)/sqrt(2p(1-p))\) (Patterson 2006), where p is the MAF,

"y-aware": center columns and multiply by beta of univariate linear regression. See https://goo.gl/8G8WMa.

thr.eigval

Threshold to remove "unsignificant" PCs. Default is 1e-3.

Value

A matrix of PCs.

Details

See BigXYt.

Note that for the Eigen decomposition, only R is used because is faster (see http://goo.gl/UYJcCw). If you want a large number of eigenvectors/values, you should really considerer using Microsoft R Open for speed.

References

Patterson N, Price AL, Reich D (2006) Population Structure and Eigenanalysis. PLoS Genet 2(12): e190. doi: 10.1371/journal.pgen.0020190

See also

bigSNP prcomp

Examples

## Not run: ------------------------------------
# 
#   # Creating directory for backing files
#   if (!dir.exists("backingfiles")) dir.create("backingfiles")
# 
#   if (!file.exists("backingfiles/test_doc.bk")) {
#     # Reading the bedfile and storing the data
#     bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
#     test <- BedToBig(bedfile, 50, "test_doc")
#   } else {
#     # Loading it from backing files
#     test <- AttachBigSNP("test_doc")
#   }
# 
#   # Get the first 10 PCs
#   test2 <- PCA.bigSNP(test, block.size = 1000, k = 10)
# 
#   # Make a nice plot out of the two first PCs
#   grp.pop <- test$fam$family.ID
#   grp.pop[grp.pop %in% paste0("POP", 1:5)] <- 1
#   grp.pop[grp.pop %in% paste0("POP", 6:11)] <- 2
#   grp.pop[grp.pop %in% paste0("POP", 12:19)] <- 3
#   x <- test2[, 1]
#   is.right = (max(x) > max(-x))
#   y <- test2[, 2]
#   is.top = (max(y) > max(-y))
#   legend.pos <- ifelse(is.top, ifelse(is.right, "topright", "topleft"),
#                        ifelse(is.right, "bottomright", "bottomleft"))
#   plot(x, y, xlab = "PC1", ylab = "PC2", col = grp.pop)
#   legend(legend.pos, legend = c("POP1 to POP5",
#                                 "POP6 to POP11",
#                                 "POP12 to POP19"),
#          col = 1:3, pch = 1, cex = 0.8)
# 
#   # Train PCs with only half the data
#   n <- nrow(test$genotypes)
#   ind.train <- sort(sample(n, n/2))
#   test3 <- PCA.bigSNP(test, block.size = 1000, k = 10,
#                       ind.train = ind.train)
# 
#   # Compare with prcomp
#   test4 <- prcomp(test$genotypes[,], scale. = TRUE)
#   plot(test4$x, col = grp.pop)
## ---------------------------------------------