AUC

Compute the Area Under the ROC Curve (AUC) of a predictor and possibly its 95% confidence interval.

snp_aucSample(pred, target, nsim = 1e+07, seed = NULL, digits = 3)

snp_aucSampleConf(pred, target, nboot = 10000, nsim = 10000, seed = NULL,
  digits = 3)

Arguments

pred

Vector of predictions.

target

Vector of true labels (must have exactly two levels, no missing values).

nsim

Number of pairs (case/control) that are compared in order to approximate the probability.

seed

See set.seed. Use it for reproducibility.

digits

See round.

nboot

Number of bootstrap samples to evaluate the 95% CI.

Value

The AUC, a probability, and possibly its 2.5% and 97.5% quantiles (95% CI).

Details

Other packages provide ways to compute the AUC. I chose to compute the AUC through its statistical definition as a probability: $$P(score(x_{case}) > score(x_{control}))$$. Note that I consider equality between scores as a 50 one being greater than the other.

References

Tom Fawcett. 2006. An introduction to ROC analysis. Pattern Recogn. Lett. 27, 8 (June 2006), 861-874. http://dx.doi.org/10.1016/j.patrec.2005.10.010.

See also

pROC::auc AUC::auc

Examples

set.seed(1)

snp_aucSample(c(0, 0), 0:1) # Equality of scores

#> [1] 0.5

snp_aucSample(c(0.2, 0.1, 1), c(-1, -1, 1)) # Perfect AUC

#> [1] 1

x <- rnorm(100)
z <- rnorm(length(x), x, abs(x))
y <- sign(z)
print(snp_aucSample(x, y))

#> [1] 0.825

print(snp_aucSampleConf(x, y, nboot = 1e4, nsim = 1e3))

#>  Mean  2.5% 97.5%    Sd 
#> 0.825 0.728 0.911 0.047