Tcrossprod

Compute \(X.row X.row^T\) for a Filebacked Big Matrix X after applying a particular scaling to it.

bed_tcrossprodSelf(
  obj.bed,
  fun.scaling = bed_scaleBinom,
  ind.row = rows_along(obj.bed),
  ind.col = cols_along(obj.bed),
  block.size = block_size(length(ind.row))
)

Arguments

obj.bed
Object of type bed, which is the mapping of some bed file. Use obj.bed <- bed(bedfile) to get this object.

fun.scaling
A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}.$$ Default doesn't use any scaling.

ind.row
An optional vector of the row indices that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices that are used. If not specified, all columns are used. Don't use negative indices.

block.size
Maximum number of columns read at once. Default uses block_size.

Value

A temporary FBM, with the following two attributes:

a numeric vector center of column scaling,

a numeric vector scale of column scaling.

Examples

bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
obj.bed <- bed(bedfile)

K <- bed_tcrossprodSelf(obj.bed)
K[1:4, 1:6] / ncol(obj.bed)

#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#> [1,] 1.09900885 0.05531637 0.08230576 0.06440704 0.06389930 0.06632927
#> [2,] 0.05531637 1.14651667 0.08627960 0.09470694 0.12318884 0.09608808
#> [3,] 0.08230576 0.08627960 1.11142319 0.06860068 0.09951989 0.04400732
#> [4,] 0.06440704 0.09470694 0.06860068 1.06728731 0.08570288 0.09599564