bigsnpr: Analysis of Massive SNP Arrays

Easy-to-use, efficient, flexible and scalable tools for analyzing massive SNP arrays. Privé et al. (2018) doi:10.1093/bioinformatics/bty185 .

Arguments

G

A FBM.code256 (typically <bigSNP>$genotypes).
You shouldn't have missing values. Also, remember to do quality control, e.g. some algorithms in this package won't work if you use SNPs with 0 MAF.

Gna

A FBM.code256 (typically <bigSNP>$genotypes).
You can have missing values in these data.

x

A bigSNP.

infos.chr

Vector of integers specifying each SNP's chromosome.
Typically <bigSNP>$map$chromosome.

infos.pos

Vector of integers specifying the physical position on a chromosome (in base pairs) of each SNP.
Typically <bigSNP>$map$physical.pos.

nploidy

Number of trials, parameter of the binomial distribution. Default is 2, which corresponds to diploidy, such as for the human genome.

ind.row

An optional vector of the row indices (individuals) that are used. If not specified, all rows are used.
Don't use negative indices.

ind.col

An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used.
Don't use negative indices.

ncores

Number of cores used. Default doesn't use parallelism. You may use nb_cores().

is.size.in.bp

Deprecated.

obj.bed

Object of type bed, which is the mapping of some bed file. Use obj.bed <- bed(bedfile) to get this object.

See also

Author

Maintainer: Florian Privé florian.prive.21@gmail.com

Other contributors:

Michael Blum [thesis advisor]

Hugues Aschard hugues.aschard@pasteur.fr [thesis advisor]

Bjarni Jóhann Vilhjálmsson bjv@econ.au.dk [thesis advisor]