Modify a "bigSNP" and save these modifications.

Modify a bigSNP and save these modifications in the corresponding .rds file. As bigSNP is an S3 class, you can add any slot you want to this class, then use SaveModifs to save modifications in the corresponding .rds backing file.

GetPops: Get populations of individuals by matching from an external file or get them directly from the slot "family.ID" of the slot "fam". It will add an slot called "pop" to the slot "fam".

GetPhenos: Modify phenotypes to be -1 (unaffected), 1 (affected) or NA (missing). It will add an slot called "pheno" to the slot "fam".

SaveModifs(x)

GetPops(x, pop.files = NULL, col.sample.ID, col.family.ID, ...)

GetPhenos(x, coded01 = FALSE)

Arguments

x

A bigSNP.

pop.files

Character vector of file names where to find the population of the individuals of the study. Default just copy the slot "family.ID" to the slot "pop".

col.sample.ID

Index of the column containing the sample IDs to match with those of the study.

col.family.ID

Index of the column containt the populations.

...

Any additional parameter to pass to fread.

coded01

Are the phenotypes coded 0 (unaffected) / 1 (affected) rather than respectively 1 and 2? Default is FALSE.

Value

The (modified) bigSNP.

See also

rle

fread

Examples

# Creating directory for backing files
if (!dir.exists("backingfiles")) dir.create("backingfiles")

if (!file.exists("backingfiles/test_doc.bk")) {
  # Reading the bedfile and storing the data
  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
  test <- BedToBig(bedfile, 50, "test_doc")
} else {
  # Loading it from backing files
  test <- AttachBigSNP("test_doc")
}

# Just after reading
print(rle(test$fam$family.ID))

#> Run Length Encoding
#>   lengths: int [1:19] 34 28 15 35 31 25 30 35 35 20 ...
#>   values : chr [1:19] "POP1" "POP2" "POP3" "POP4" "POP5" "POP6" "POP7" ...

print(test$fam$pop) # NULL

#>   [1] "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1" 
#>  [10] "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1" 
#>  [19] "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1" 
#>  [28] "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP1"  "POP2"  "POP2" 
#>  [37] "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2" 
#>  [46] "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2" 
#>  [55] "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP2"  "POP3" 
#>  [64] "POP3"  "POP3"  "POP3"  "POP3"  "POP3"  "POP3"  "POP3"  "POP3"  "POP3" 
#>  [73] "POP3"  "POP3"  "POP3"  "POP3"  "POP3"  "POP4"  "POP4"  "POP4"  "POP4" 
#>  [82] "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4" 
#>  [91] "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4" 
#> [100] "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4"  "POP4" 
#> [109] "POP4"  "POP4"  "POP4"  "POP4"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5" 
#> [118] "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5" 
#> [127] "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5" 
#> [136] "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP5"  "POP6" 
#> [145] "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6" 
#> [154] "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6" 
#> [163] "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP6"  "POP7"  "POP7"  "POP7" 
#> [172] "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7" 
#> [181] "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7" 
#> [190] "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7"  "POP7" 
#> [199] "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8" 
#> [208] "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8" 
#> [217] "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8" 
#> [226] "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP8"  "POP9" 
#> [235] "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9" 
#> [244] "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9" 
#> [253] "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9" 
#> [262] "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP9"  "POP10" "POP10"
#> [271] "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10"
#> [280] "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10" "POP10"
#> [289] "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11"
#> [298] "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11" "POP11"
#> [307] "POP11" "POP11" "POP11" "POP11" "POP12" "POP12" "POP12" "POP12" "POP12"
#> [316] "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12"
#> [325] "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12" "POP12"
#> [334] "POP12" "POP12" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13"
#> [343] "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13"
#> [352] "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP13"
#> [361] "POP13" "POP13" "POP13" "POP13" "POP13" "POP13" "POP14" "POP14" "POP14"
#> [370] "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14"
#> [379] "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14" "POP14"
#> [388] "POP14" "POP14" "POP14" "POP14" "POP15" "POP15" "POP15" "POP15" "POP15"
#> [397] "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15"
#> [406] "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15"
#> [415] "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15" "POP15"
#> [424] "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16"
#> [433] "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16"
#> [442] "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP16"
#> [451] "POP16" "POP16" "POP16" "POP16" "POP16" "POP16" "POP17" "POP17" "POP17"
#> [460] "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17"
#> [469] "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17" "POP17"
#> [478] "POP17" "POP17" "POP17" "POP17" "POP17" "POP18" "POP18" "POP18" "POP18"
#> [487] "POP18" "POP18" "POP18" "POP18" "POP18" "POP18" "POP18" "POP18" "POP18"
#> [496] "POP18" "POP18" "POP18" "POP19" "POP19" "POP19" "POP19" "POP19" "POP19"
#> [505] "POP19" "POP19" "POP19" "POP19" "POP19" "POP19" "POP19" "POP19" "POP19"
#> [514] "POP19" "POP19" "POP19" "POP19"

# Get populations from slot "family.ID"
test <- GetPops(test)
print(rle(test$fam$pop))

#> Run Length Encoding
#>   lengths: int [1:19] 34 28 15 35 31 25 30 35 35 20 ...
#>   values : chr [1:19] "POP1" "POP2" "POP3" "POP4" "POP5" "POP6" "POP7" ...

# Get populations clusters from external files
files <- system.file("extdata", paste0("cluster", 1:3), package = "bigsnpr")
test <- GetPops(x = test, pop.files = files, col.sample.ID = 2, col.family.ID = 3)

#> Error in GetPops(x = test, pop.files = files, col.sample.ID = 2, col.family.ID = 3): impossible de trouver la fonction "foreach2"

print(rle(test$fam$pop))

#> Run Length Encoding
#>   lengths: int [1:19] 34 28 15 35 31 25 30 35 35 20 ...
#>   values : chr [1:19] "POP1" "POP2" "POP3" "POP4" "POP5" "POP6" "POP7" ...