LD clumping

LD clumping (and thresholding) for a bigSNP.

Arguments

fun.stats

A function that takes a big.matrix X and ind.train as parameters and returns a named list of S and pS for every column, which are statistics and associated p-values. pS needs to be computed through a decreasing function of S. For example, if S follows the standard normal distribution, you should use abs(S) instead and compute pS = 2 * pnorm(abs(S), lower.tail = FALSE).

size

Radius of the window's size for the LD evaluations. Default is 1000, which seems pretty conservative for a standard chip of less than a million SNPs.

thr.pvalue

Threshold on \(-log_{10}(p-value)\) to assess which SNPs are kept. A default of 1 is very conservative and has the purpose to accelerate computations.

thr.corr

Threshold on the correlation between two SNPs. SNPs which are too correlated with another SNP which is more correlated with the disease are pruned.

exclude

Vector of indices of SNPs to exclude anyway. For example, can be used to exclude long-range LD regions (see Price2008)

References

Price AL, Weale ME, Patterson N, et al. Long-Range LD Can Confound Genome Scans in Admixed Populations. Am J Hum Genet. 2008;83(1):132-135. http://dx.doi.org/10.1016/j.ajhg.2008.06.005.

Examples

# Creating directory for backing files
if (!dir.exists("backingfiles")) dir.create("backingfiles")

if (!file.exists("backingfiles/test_doc.bk")) {
  # Reading the bedfile and storing the data
  bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
  test <- BedToBig(bedfile, 50, "test_doc")
} else {
  # Loading it from backing files
  test <- AttachBigSNP("test_doc")
}

# generating random phenotypes
test$fam$pheno <- sample(c(-1, 1), size = nrow(test$fam),
                         replace = TRUE)

ind <- Prune(test)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "Prune"

# ind <- Prune(test, ncores = 2)

R2 <- RsqClass(test$genotypes, test$fam$pheno)

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "RsqClass"

par.save <- par(mfrow = c(1, 2))

plot(R2, type = 'h')

#> Error in plot(R2, type = "h"): objet 'R2' introuvable

plot(ind, R2[ind], type = "h")

#> Error in plot(ind, R2[ind], type = "h"): objet 'ind' introuvable

par(par.save)