LD pruning and clumping

For a bigSNP:

snp_pruning: LD pruning. Similar to "--indep-pairwise size 1 thr.corr" in http://pngu.mgh.harvard.edu/~purcell/plink/summary.shtml#prune (step is fixed to 1).

snp_clumping: LD clumping.

snp_indLRLDR: Get SNP indices of long-range LD regions.

snp_clumping(x, S, ind.train = seq(nrow(X)), size = 500, thr.corr = 0.5,
  exclude = NULL, ncores = 1)

snp_pruning(x, ind.train = seq(nrow(X)), size = 50, thr.corr = 0.5,
  exclude = NULL, ncores = 1)

snp_indLRLDR(x, LD.regions = LD.wiki34)

Arguments

x

A bigSNP.

S

A vector of column statistics which express the importance of each SNP (the more important is the SNP, the greater should be the corresponding statistic). For example, if S follows the standard normal distribution, you should probably use abs(S) instead.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

size

This parameter should be adjusted with respect to the number of SNPs.

for clumping: Radius of the window's size for the LD evaluations. Default is 500 (I use this for a chip of 500K SNPs).

for pruning: Diameter of the window's size for the LD evaluations. Default is 50 (as in PLINK).

thr.corr

Threshold on the correlation between two SNPs. Default is 0.5.

exclude

Vector of indices of SNPs to exclude anyway. For example, can be used to exclude long-range LD regions (see Price2008). Another use can be for thresholding with respect to p-values associated with S.

ncores

Number of cores used. Default doesn't use parallelism.

LD.regions

A data.frame with columns "Chr", "Start" and "Stop". Default use the table of 34 long-range LD regions that you can find https://goo.gl/0Ou7uI.

Value

snp_pruning & snp_pruning: SNP indices which are kept.

snp_indLRLDR: SNP indices to be used as (part of) the exclude parameter of snp_pruning or snp_clumping.

Details

I recommend to use clumping rather than pruning. See https://privefl.github.io/bigsnpr/articles/pruning-vs-clumping.html.

References

Price AL, Weale ME, Patterson N, et al. Long-Range LD Can Confound Genome Scans in Admixed Populations. Am J Hum Genet. 2008;83(1):132-135. http://dx.doi.org/10.1016/j.ajhg.2008.06.005