Read PLINK files into a "bigSNP".

Functions to read bed/bim/fam files into a bigSNP. For more information on these formats, please visit http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#bed. For other formats, please use PLINK to convert them in bedfiles, which require minimal space to store and are faster to read.

BedToBig(bedfile, block.size = 1000, backingfile,
  backingpath = "backingfiles")

AttachBigSNP(backingfile, backingpath = "backingfiles", readonly = TRUE)

snp_readExample(backingfile = "test_doc", backingpath = "backingfiles")

snp_readHapMap3(backingfile = "test_HapMap3", backingpath = "backingfiles")

Arguments

bedfile

Path to file with extension .bed. You need the corresponding .bim and .fam in the same directory.

backingfile

The root name for the backing file(s) for the cache of the resulting object.

backingpath

The path to the directory containing the file backing cache. Default is "backingfiles". It needs to exist (use dir.create).

readonly

Is the big.matrix read only? Default is TRUE.

Value

A bigSNP. Reading PLINK files creates backingfile.bk, backingfile.desc and backingfile.rds in directory backingpath. You shouldn't read from PLINK files more than once. Instead, use AttachBigSNP to load this object in another session from backing files.

Note

The implementation was originally inspired from the code of https://github.com/andrewparkermorgan/argyle. I did many optimizations. Especially, online reading into a big.matrix makes it memory-efficient.

Examples

## Not run: ------------------------------------
# 
#   bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
# 
#   # Creating directory for backing files
#   if (!dir.exists("backingfiles")) dir.create("backingfiles")
# 
#   # Reading the bedfile and storing the data in directory "backingfiles"
#   if (!file.exists("backingfiles/test_doc.bk"))
#     test <- BedToBig(bedfile, 50, "test_doc")
# 
#   # Removing the R object
#   rm(test)
# 
#   # Loading it from backing files
#   test <- AttachBigSNP("test_doc")
# 
#   str(test)
#   print(dim(test$genotypes))
#   print(test$genotypes[1:8, 1:8])
## ---------------------------------------------