Determine reference divergence

Determine reference divergence while accounting for strand flips. This does not remove ambiguous alleles.

same_ref(ref1, alt1, ref2, alt2)

Arguments

ref1
The reference alleles of the first dataset.

alt1
The alternative alleles of the first dataset.

ref2
The reference alleles of the second dataset.

alt2
The alternative alleles of the second dataset.

Value

A logical vector whether the references alleles are the same. Missing values can result from missing values in the inputs or from ambiguous strands.

See also

Examples

same_ref(ref1 = c("A", "C", "T", "G", NA),
         alt1 = c("C", "T", "C", "A", "A"),
         ref2 = c("A", "C", "A", "A", "C"),
         alt2 = c("C", "G", "G", "G", "A"))

#> [1]  TRUE    NA  TRUE FALSE    NA