MAF

Minor Allele Frequency.

snp_MAF(G, ind.row = rows_along(G), ind.col = cols_along(G),
  nploidy = getOption("bigsnpr.nploidy"))

Arguments

G
A FBM.code256 (typically <bigSNP>$genotypes). You shouldn't have missing values in your data or SNPs with 0 MAF.

ind.row
An optional vector of the row indices (individuals) that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used. Don't use negative indices.

nploidy
Number of trials, parameter of the binomial distribution. Default is 2, which corresponds to diploidy, such as for the human genome.

Value

A vector of MAFs, corresponding to ind.col.

Examples

obj.bigsnp <- snp_attachExtdata()
str(maf <- snp_MAF(obj.bigsnp$genotypes))

#>  num [1:4542] 0.342 0.206 0.237 0.185 0.456 ...