MAX3 statistic

Counts the number of 0, 1, 2 by SNP and phenotype (cases/controls) and then compute the MAX3 statistic.

snp_MAX3(x, ind.train = rows_along(X.), val = c(0, 0.5, 1))

Arguments

x

A bigSNP.

val

Computing \(\smash{\displaystyle\max_{x \in val}}~Z_{CATT}^2(x)\). Default is c(0, 0.5, 1) and corresponds to the MAX3 statistic. Only c(0, 1) corresponds to MAX2. And only 0.5 corresponds to the Armitage trend test. Finally, seq(0, 1, length.out = L) corresponds to MAXL.

Value

A data.frame of S and pS for every column, which are MAX3 statistics and associated p-values. P-values are in fact the minimum of the p-values of each test separately (so they are biased downward). One can use genomic control to rescale these p-values.

References

Zheng, G., Yang, Y., Zhu, X., & Elston, R. (2012). Robust Procedures. Analysis Of Genetic Association Studies, 151-206. http://dx.doi.org/10.1007/978-1-4614-2245-7_6.

Examples

set.seed(1)

# constructing a fake genotype big.matrix
a <- big.matrix(10, 12, type = "char", shared = FALSE)
a[] <- sample(c(0, 1, 2), 150, TRUE)

#> Assignment will down cast from double to char
#> Hint: To remove this warning type:  options(bigmemory.typecast.warning=FALSE)

#> Warning: 

#> Error in SetAll.bm(x, value): number of items to replace is not a multiple of replacement length

print(a[,])

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#>  [1,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [2,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [3,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [4,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [5,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [6,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [7,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [8,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#>  [9,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [10,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$affection <- c(rep(1, 5), rep(-1, 5))

# Get MAX3 statistics
print(snp_MAX3(fake))

#> Error: 'X.code' is not of class 'BM.code' (or associated descriptor).