PRS

Polygenic Risk Scores with possible clumping and thresholding.

snp_PRS(x, betas, ind.test, lpS = NULL, thr.list = 0, ...)

Arguments

x

A bigSNP.

betas

Numeric vector of weights associated with each SNP. You may want to see big_univRegLin or big_univRegLog.

ind.test

The individuals on whom to project the scores.

lpS

Numeric vector of -log10(p.value) associated with betas. Default doesn't use thresholding.

thr.list

Threshold vector on lpS at which SNPs are excluded if they are not significant enough. Default doesn't use thresholding.

...

Arguments passed on to snp_clumping

S

A vector of column statistics which express the importance of each SNP (the more important is the SNP, the greater should be the corresponding statistic). For example, if S follows the standard normal distribution, you should probably use abs(S) instead.

size

This parameter should be adjusted with respect to the number of SNPs.

for clumping: Radius of the window's size for the LD evaluations. Default is 1000 (I use this for a chip of 500K SNPs).

for pruning: Diameter of the window's size for the LD evaluations. Default is 50 (as in PLINK 1.07).

thr.r2

Threshold over the squared correlation between two SNPs. Default is 0.5.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all rows are used.

ncores

Number of cores used. Default doesn't use parallelism.

Value

A matrix of scores, where rows correspond to ind.test and columns correspond to thr.list.