Attach a "bigSNP" from backing files

Load a bigSNP from backing files into R.

snp_attach(rdsfile)

Arguments

rdsfile

The path of the ".rds" which stores the bigSNP object.

Value

The bigSNP object.

Details

This is often just a call to readRDS. But it also checks if you have moved the two (".bk" and ".rds") backing files to another directory.

Examples

bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
print(bedfile)

#> [1] "D:/Projets/bigsnpr/inst/extdata/example.bed"

# Reading the bedfile and storing the data in directory "backingfiles"
PATH <- "backingfiles/test_doc.bk"
if (!file.exists(PATH))
  PATH <- snp_readBed(bedfile, backingfile = "test_doc")

#> Creating directory "backingfiles" which didn't exist..

# Loading the data from backing files
test <- snp_attach(PATH)

str(test)

#> List of 4
#>  $ genotypes:Formal class 'BM.code.descriptor' [package "bigstatsr"] with 2 slots
#>   .. ..@ code       : num [1:256] 0 1 2 NA NA NA NA NA NA NA ...
#>   .. ..@ description:List of 13
#>   .. .. ..$ sharedType: chr "FileBacked"
#>   .. .. ..$ filename  : chr "test_doc.bk"
#>   .. .. ..$ dirname   : chr "backingfiles/"
#>   .. .. ..$ totalRows : int 517
#>   .. .. ..$ totalCols : int 4542
#>   .. .. ..$ rowOffset : num [1:2] 0 517
#>   .. .. ..$ colOffset : num [1:2] 0 4542
#>   .. .. ..$ nrow      : num 517
#>   .. .. ..$ ncol      : num 4542
#>   .. .. ..$ rowNames  : NULL
#>   .. .. ..$ colNames  : NULL
#>   .. .. ..$ type      : chr "raw"
#>   .. .. ..$ separated : logi FALSE
#>  $ fam      :'data.frame':	517 obs. of  6 variables:
#>   ..$ family.ID  : chr [1:517] "POP1" "POP1" "POP1" "POP1" ...
#>   ..$ sample.ID  : chr [1:517] "IND0" "IND1" "IND2" "IND3" ...
#>   ..$ paternal.ID: int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ maternal.ID: int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ sex        : int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ affection  : int [1:517] -9 -9 -9 -9 -9 -9 -9 -9 -9 -9 ...
#>  $ map      :'data.frame':	4542 obs. of  6 variables:
#>   ..$ chromosome  : int [1:4542] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ marker.ID   : chr [1:4542] "SNP0" "SNP1" "SNP2" "SNP3" ...
#>   ..$ genetic.dist: int [1:4542] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ physical.pos: int [1:4542] 0 1000 2000 3000 4000 5000 6000 7000 8000 9000 ...
#>   ..$ allele1     : chr [1:4542] "A" "T" "T" "T" ...
#>   ..$ allele2     : chr [1:4542] "T" "A" "A" "A" ...
#>  $ savedIn  : chr "backingfiles/test_doc.rds"
#>  - attr(*, "class")= chr "bigSNP"

print(dim(test$genotypes))

#> [1]  517 4542

print(test$genotypes[1:8, 1:8])

#> Error in test$genotypes[1:8, 1:8]: object of type 'S4' is not subsettable