Imputation

Imputation using Beagle version 4.

snp_beagleImpute(beagle.path, plink.path, bedfile.in, bedfile.out = NULL,
  memory.max = 3, ncores = 1, extra.options = "")

Arguments

beagle.path

Path to the executable of Beagle v4+.

plink.path

Path to the executable of PLINK 1.9.

bedfile.in

Path to the input bedfile.

bedfile.out

Path to the output bedfile. Default is created by appending "_impute" to prefix.in (bedfile.in without extension).

memory.max

Max memory (in Gb) to be used. It is internally rounded to be an integer. Default is 3.

ncores

Number of cores to be used. Default is 1. An usually good value for this parameter is ncores = parallel::detectCores() - 1.

extra.options

Other options to be passed to Beagle as a string. More options can be found at Beagle's website.

Value

The path of the new bedfile.

Details

Downloads and more informations can be found at the following websites

https://www.cog-genomics.org/plink2,

https://faculty.washington.edu/browning/beagle/beagle.html.

References

B L Browning and S R Browning (2016). Genotype imputation with millions of reference samples. Am J Hum Genet 98:116-126. dx.doi.org/doi:10.1016/j.ajhg.2015.11.020