Truncated SVD with pruning

Fast truncated SVD which iteratively try to remove long-range LD regions which appear in loadings of SVD.

snp_clumpedSVD(G, infos.chr, ind.row = rows_along(G),
  ind.col = cols_along(G), fun.scaling = snp_scaleBinom(),
  thr.r2.init = 0.2, thr.r2.step = 2, size = 500, k = 10,
  roll.size = 50, int.min.size = 20, ncores = 1, verbose = TRUE)

Arguments

G

A BM.code.descriptor (typically <bigSNP>$genotypes). You shouldn't have missing values in your data or SNPs with 0 MAF.

infos.chr

Vector of integers specifying each SNP's chromosome. Typically <bigSNP>$map$chromosome.

ind.row

An optional vector of the row indices (individuals) that are used. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used. Don't use negative indices.

fun.scaling

A function that returns a named list of mean and sd for every column, to scale each of their elements such as followed: $$\frac{X_{i,j} - mean_j}{sd_j}.$$ Default is snp_scaleBinom().

thr.r2.init

Threshold on squared correlation between SNPs. Default is 0.2.

thr.r2.step

Coefficient of division of thr.r2 at each step. Default is 2 which means that at iteration 2, thr.r2 = 0.1, at iteration 3, thr.r2 = 0.05, etc.

size

Radius of the window's size for the LD evaluations of the initial step of clumping. Default is 500.

k

Number of singular vectors/values to compute. Default is 10. This algorithm should be used to compute only a few singular vectors/values.

roll.size

Radius of rolling windows to smooth log-p-values.

int.min.size

Minimum size of intervals of consecutive significant indices.

ncores

Number of cores used. Default doesn't use parallelism.

verbose

Output some information on the iterations? Default is TRUE.

Value

A named list (an S3 class "big_SVD") of

d, the singular values,

u, the left singular vectors,

v, the right singular vectors,

niter, the number of the iteration of the algorithm,

nops, number of Matrix-Vector multiplications used,

means, the centering vector,

sds, the scaling vector.

Note that to obtain the Principal Components, you must use predict on the result. See examples.

Details

Improvements will come in the future, as for example, warm starts in order to make the SVD computations faster.

Examples

ex <- snp_attachExtdata()

obj.svd <- snp_clumpedSVD(G = ex$genotypes,
                          infos.chr = ex$map$chromosome)

#> First phase of clumping at r2 > 0.2.. keep 4270 SNPs.
#> 
#> Iteration 1:
#> Computing SVD..
#> 
#> Converged!

str(obj.svd)

#> List of 7
#>  $ d    : num [1:10] 235.4 148 105.5 96.4 94.9 ...
#>  $ u    : num [1:517, 1:10] 0.0801 0.0798 0.0646 0.0781 0.0818 ...
#>  $ v    : num [1:4270, 1:10] 0.00174 -0.03142 0.01527 -0.0132 -0.0154 ...
#>  $ niter: int 10
#>  $ nops : int 170
#>  $ means: num [1:4270] 1.59 1.53 1.63 1.09 1.29 ...
#>  $ sds  : num [1:4270] 0.572 0.601 0.549 0.704 0.677 ...
#>  - attr(*, "class")= chr "big_SVD"
#>  - attr(*, "subset")= int [1:4270] 2 3 4 5 6 7 8 9 10 11 ...