Counts

Counts the number of 0, 1, 2 by SNP and phenotype (cases/controls) and the number of NA by individual. The number of NA by SNP can be deduced using the relation \(n_{NA} = n - (n_0 + n_1 + n_2)\).

snp_counts(x, has.pheno = TRUE)

Arguments

x

A bigSNP.

has.pheno

Does x has well-specified phenotypes? i.e. x$fam$affection is a vector with exactly two values (and the value coding for cases is larger than the value coding for controls). So, in particular, no missing value is allowed. Default is TRUE.

Value

A list of:

one or two matrices of size 3*m (m is the number of SNPs) representing the counts of 0, 1 and 2 (by status),

an integer vector of size n (n is the number of individuals) corresponding to the number of missing values per individual.

Examples

set.seed(1)

# constructing a fake genotype big.matrix
a <- big.matrix(10, 15, type = "char")
a[] <- sample(c(0, 1, 2, NA), 150, TRUE)

#> Assignment will down cast from double to char
#> Hint: To remove this warning type:  options(bigmemory.typecast.warning=FALSE)

#> Warning: 

print(a[,])

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    1   NA    2    0    2    1   NA   NA    2     0    NA    NA     2
#>  [2,]    1    2   NA    0    1   NA    0   NA    0     0     1     2     1
#>  [3,]    0   NA    0    1    1   NA    2    1    0     2     2     2     0
#>  [4,]    1    0    1    2   NA    1    0    1    1     0     0     2     2
#>  [5,]   NA    2    2    2    0   NA    0   NA   NA     1     0    NA     0
#>  [6,]    1    1    2    1   NA   NA    0    2    2     2     2     2    NA
#>  [7,]    1   NA    1   NA    1    1    0    2   NA    NA     2     2     2
#>  [8,]    2    2   NA    1   NA    2    0    1    2     1     0     2     2
#>  [9,]    1   NA    1    1    1    1    2    1    1     1     0     0     1
#> [10,]    0    2    0    1    1    1   NA   NA    1     0     2     1     1
#>       [,14] [,15]
#>  [1,]     2    NA
#>  [2,]     0     1
#>  [3,]     2     0
#>  [4,]     0     1
#>  [5,]     1     2
#>  [6,]     0     1
#>  [7,]     1     2
#>  [8,]    NA    NA
#>  [9,]     1    NA
#> [10,]    NA     1

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$affection <- c(rep(2, 5), rep(1, 5))

# Get counts
print(test1 <- snp_counts(fake))

#> $cols.cases
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    1    1    2    1    0    3    0    2     3     2     0     2     2
#> [2,]    3    0    1    1    2    2    0    2    1     1     1     0     1     1
#> [3,]    0    2    2    2    1    0    1    0    1     1     1     3     2     2
#>      [,15]
#> [1,]     1
#> [2,]     2
#> [3,]     1
#> 
#> $cols.controls
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    1    0    1    0    0    0    3    0    0     1     2     1     0     1
#> [2,]    3    1    2    4    3    3    0    2    2     2     0     1     2     2
#> [3,]    1    2    1    0    0    1    1    2    2     1     3     3     2     0
#>      [,15]
#> [1,]     0
#> [2,]     2
#> [3,]     1
#> 
#> $rows
#>  [1] 6 3 2 1 5 3 4 4 2 3
#> 

print(test2 <- snp_counts(fake, has.pheno = FALSE))

#> $cols
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,]    2    1    2    2    1    0    6    0    2     4     4     1     2     3
#> [2,]    6    1    3    5    5    5    0    4    3     3     1     1     3     3
#> [3,]    1    4    3    2    1    1    2    2    3     2     4     6     4     2
#>      [,15]
#> [1,]     1
#> [2,]     4
#> [3,]     2
#> 
#> $rows
#>  [1] 6 3 2 1 5 3 4 4 2 3
#> 

# Same results
test1$cols <- test1$cols.controls + test1$cols.cases
print(all.equal(test1$cols, test2$cols))

#> [1] TRUE

print(all.equal(test1$rows, test2$rows))

#> [1] TRUE