Double a genotype matrix

Recode each column of a genotype matrix to be two columns, one encoding a recessive status and the other a dominant status. Values will be rounded so that you have only 0s, 1s and 2s in your data.

snp_double(x)

Arguments

x

A bigSNP.

Value

The new "doubled" genotype matrix, as a BM.code.descriptor, which values are 0s and 1s.

Examples

test <- snp_attachExtdata()
X <- attach.BM(test$genotypes)
X[1:8, 1:5]

#>      [,1] [,2] [,3] [,4] [,5]
#> [1,]    2    2    0    2    1
#> [2,]    1    2    1    2    2
#> [3,]    2    1    1    2    0
#> [4,]    2    2    0    2    0
#> [5,]    1    2    2    2    0
#> [6,]    2    1    2    2    0
#> [7,]    2    1    1    2    0
#> [8,]    1    2    1    1    1

rm(X)

test2 <- snp_double(test)
attach.BM(test2)[1:8, 1:10]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]    1    1    1    1    0    0    1    1    0     1
#> [2,]    0    1    1    1    0    1    1    1    1     1
#> [3,]    1    1    0    1    0    1    1    1    0     0
#> [4,]    1    1    1    1    0    0    1    1    0     0
#> [5,]    0    1    1    1    1    1    1    1    0     0
#> [6,]    1    1    0    1    1    1    1    1    0     0
#> [7,]    1    1    0    1    0    1    1    1    0     0
#> [8,]    0    1    1    1    0    1    0    1    0     1