Get population infos

Get populations of individuals by matching from an external file and overwrite the slot family.ID of the slot fam.

snp_getPops(x, pop.files, col.sample.ID, col.family.ID, ...)

Arguments

pop.files

Character vector of file names where to find the population of the individuals of the study.

col.sample.ID

Index of the column containing the sample IDs to match with those of the study.

col.family.ID

Index of the column containt the populations.

...

Any additional parameter to pass to fread.

Value

The modified bigSNP (modifications are saved in .rds).

Examples

test <- snp_readExample()

# Just after reading
print(rle(test$fam$family.ID))

#> Run Length Encoding
#>   lengths: int [1:19] 34 28 15 35 31 25 30 35 35 20 ...
#>   values : chr [1:19] "POP1" "POP2" "POP3" "POP4" "POP5" "POP6" "POP7" ...

# Get populations clusters from external files
files <- system.file("extdata", paste0("cluster", 1:3), package = "bigsnpr")
print(data.table::fread(files[1]))

#>       V1     V2      V3
#>   1:   1   IND0 POP-1-5
#>   2:   2   IND1 POP-1-5
#>   3:   3   IND2 POP-1-5
#>   4:   4   IND3 POP-1-5
#>   5:   5   IND4 POP-1-5
#>  ---                   
#> 139: 139 IND138 POP-1-5
#> 140: 140 IND139 POP-1-5
#> 141: 141 IND140 POP-1-5
#> 142: 142 IND141 POP-1-5
#> 143: 143 IND142 POP-1-5

test <- snp_getPops(test, pop.files = files, col.sample.ID = 2,
                    col.family.ID = 3)
print(rle(test$fam$family.ID))

#> Run Length Encoding
#>   lengths: int [1:3] 143 167 207
#>   values : chr [1:3] "POP-1-5" "POP-6-11" "POP-12-19"