Manhattan plot

Creates a manhattan plot.

snp_manhattan(gwas, infos.chr, infos.pos, colors = c("black", "grey60"),
  dist.sep.chrs = 1e+07, ind.highlight = integer(0),
  col.highlight = "red", labels = all.chr, npoints = NULL, coeff = 1)

Arguments

gwas

A mhtest object with the p-values associated with each SNP. Typically, the output of big_univLinReg, big_univLogReg or snp_pcadapt.

colors

Colors that used for each chromosome (they are recycled). Default is an alternance of black and grey.

dist.sep.chrs

"Physical" distance that separates two chromosomes. Default is 10 Mbp.

ind.highlight

Indices of SNPs you want to highlight (of interest). Default doesn't highlight any SNPs.

col.highlight

Color used for highlighting SNPs. Default uses red.

labels

Labels of the x axis. Default uses the number of the chromosome there are in infos.chr(sort(unique(infos.chr))). This may be useful to restrict the number of labels so that they are not overlapping.

npoints

Number of points to keep (ranked by p-value) in order to get a lighter object (and plot). Default doesn't cut anything. If used, the resulting object will have an attribute called subset giving the indices of the kept points.

coeff

Relative size of text. Default is 1.

Value

A ggplot2 object. You can plot it using the print method. You can modify it as you wish by adding layers. You might want to read http://r4ds.had.co.nz/data-visualisation.html to get more familiar with the package ggplot2.

Details

If you don't have infos of chromosome and position, you should simply use plot instead.