Identity-by-descent

Quality Control based on Identity-by-descent (IBD) computed by https://www.cog-genomics.org/plink2 using its method-of-moments.

snp_plinkIBDQC(plink.path, bedfile.in, bedfile.out = NULL, pi.hat = 0.08,
  ncores = 1, pruning.args = c(100, 0.2), do.blind.QC = TRUE,
  extra.options = "")

Arguments

plink.path

Path to the executable of PLINK 1.9.

bedfile.in

Path to the input bedfile.

bedfile.out

Path to the output bedfile. Default is created by appending "_norel" to prefix.in (bedfile.in without extension).

pi.hat

PI_HAT value threshold for individuals (first by pairs) to be excluded. Default is 0.08.

ncores

Number of cores to be used. Default is 1. An usually good value for this parameter is ncores = parallel::detectCores() - 1.

pruning.args

A vector of 2 pruning parameters, respectively the window size (in variant count) and the pairwise $r^2$ threshold (the step size is fixed to 1). Default is c(100, 0.2).

do.blind.QC

Whether to do QC with pi.hat without visual inspection. Default is TRUE. If FALSE, return the data.frame of the corresponding ".genome" file without doing QC. One could use ggplot2::qplot(Z1, Z2, data = mydf, col = RT) for visual inspection.

extra.options

Other options to be passed to PLINK as a string (for the IBD part). More options can be found at https://www.cog-genomics.org/plink/1.9/ibd.

Value

The path of the new bedfile. If no sample is filter, no new bed/bim/fam files are created and then the path of the input bedfile is returned.

References

Purcell, Shaun, Benjamin Neale, Kathe Todd-Brown, Lori Thomas, Manuel A R Ferreira, David Bender, Julian Maller, et al. 2007. PLINK: a tool set for whole-genome association and population-based linkage analyses. American Journal of Human Genetics 81 (3). Elsevier: 559–75. http://dx.doi.org/10.1086/519795.

Chang, Christopher C, Carson C Chow, Laurent CAM Tellier, Shashaank Vattikuti, Shaun M Purcell, and James J Lee. 2015. Second-generation PLINK: rising to the challenge of larger and richer datasets. GigaScience 4 (1): 7. http://dx.doi.org/10.1186/s13742-015-0047-8.

See also

snp_plinkQC

Examples

## Not run: ------------------------------------
# # On MY computer, I can do:
# plink <- "../plink_linux_x86_64/plink"
# bedfile <- "../POPRES_data/POPRES_allchr_QC.bed"
# test <- snp_plinkIBDQC(plink, bedfile, ncores = 4)
## ---------------------------------------------