Q-Q plot

Creates a quantile-quantile plot from p-values from a GWAS study.

snp_qq(gwas, lambdaGC = TRUE, coeff = 1)

Arguments

gwas

A mhtest object with the p-values associated with each SNP. Typically, the output of big_univLinReg, big_univLogReg or snp_pcadapt.

lambdaGC

Add the Genomic Control coefficient as subtitle to the plot?

coeff

Relative size of text. Default is 1.

Value

A ggplot2 object. You can plot it using the print method. You can modify it as you wish by adding layers. You might want to read http://r4ds.had.co.nz/data-visualisation.html to get more familiar with the package ggplot2.

Examples

test <- snp_attachExtdata()
svd <- big_SVD(test$genotypes, snp_scaleBinom(), k = 10)

#> (2)

tmp <- snp_pcadapt(test$genotypes, svd$u[, 1:3])

#> Error in .Call("bigstatsr_linRegPcadapt", PACKAGE = "bigstatsr", BM, U,     rowInd): "bigstatsr_linRegPcadapt" not available for .Call() for package "bigstatsr"

print(p <- snp_qq(tmp))

#> Error in graphics::plot(gwas, type = "Q-Q", coeff = coeff): object 'tmp' not found

print(p2 <- p + ggplot2::aes(text = asPlotlyText(test$map)))

#> Error in print(p2 <- p + ggplot2::aes(text = asPlotlyText(test$map))): object 'p' not found

plotly::ggplotly(p2, tooltip = "text")

#> Error in plotly::ggplotly(p2, tooltip = "text"): object 'p2' not found