Read PLINK files into a "bigSNP"

Function to read bed/bim/fam files into a bigSNP.

snp_readBed(bedfile, backingfile, backingpath = "backingfiles")

Arguments

bedfile

Path to file with extension ".bed" to read. You need the corresponding ".bim" and ".fam" in the same directory.

backingfile

The root name for the backing file(s) for the cache of the bigSNP object.

backingpath

The path to the directory containing the backing files. Default is "backingfiles".

Value

The path to the RDS file that stores the bigSNP object. Note that this function creates two other files which store the big.matrix and its descriptor. You shouldn't read from PLINK files more than once. Instead, use snp_attach to load the "bigSNP" object in any R session from backing files.

Details

For more information on these formats, please visit http://pngu.mgh.harvard.edu/~purcell/plink/data.shtml#bed. For other formats, please use PLINK to convert them in bedfiles, which require minimal space to store and are faster to read. For example, to convert from a VCF file, use the --vcf option.

Examples

bedfile <- system.file("extdata", "example.bed", package = "bigsnpr")
print(bedfile)

#> [1] "D:/Projets/bigsnpr/inst/extdata/example.bed"

# Reading the bedfile and storing the data in directory "backingfiles"
PATH <- "backingfiles/test_doc.bk"
if (!file.exists(PATH))
  PATH <- snp_readBed(bedfile, backingfile = "test_doc")

# Loading the data from backing files
test <- snp_attach(PATH)

#> Error in readRDS(rdsfile): unknown input format

str(test)

#> Error in str(test): object 'test' not found

print(dim(test$genotypes))

#> Error in print(dim(test$genotypes)): object 'test' not found

print(test$genotypes[1:8, 1:8])

#> Error in print(test$genotypes[1:8, 1:8]): object 'test' not found