Save modifications for a "bigSNP"

Save a bigSNP after having made some modifications to it. As bigSNP is an S3 class, you can add any slot you want to an object of this class, then use snp_saveModifs to save these modifications in the corresponding .rds backing file.

snp_saveModifs(x)

Arguments

x

A bigSNP.

Value

The (saved) bigSNP.

Examples

set.seed(1)

#reading example
test <- snp_readExample()

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "snp_readExample"

# I can add whatever I want to a S3 class
test$map$`p-values` <- runif(nrow(test$map))

#> Error in nrow(test$map): objet 'test' introuvable

str(test$map)

#> Error in str(test$map): objet 'test' introuvable

# reading again
test <- snp_readExample()

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "snp_readExample"

str(test$map) # new slot wasn't saved

#> Error in str(test$map): objet 'test' introuvable

# save it
test$map$`p-values` <- runif(nrow(test$map))

#> Error in nrow(test$map): objet 'test' introuvable

test <- snp_saveModifs(test)

#> Error in file.path(x$backingpath, paste0(x$backingfile, ".rds")): objet 'test' introuvable

# reading again
test <- snp_readExample()

#> Error in eval(expr, envir, enclos): impossible de trouver la fonction "snp_readExample"

str(test$map) # it is saved now

#> Error in str(test$map): objet 'test' introuvable