bigsnpr: Analysis of Massive SNP Arrays

Easy-to-use, efficient, flexible and scalable tools for the analysis of massive SNP arrays. Preprint: Privé et al. (2017) <doi:10.1101/190926>.

Arguments

G
A FBM.code256 (typically <bigSNP>$genotypes).
You shouldn't have missing values in your data or SNPs with 0 MAF.

Gna
A FBM.code256 (typically <bigSNP>$genotypes).
You can have missing values in your data.

x
A bigSNP.

infos.chr
Vector of integers specifying each SNP's chromosome.
Typically <bigSNP>$map$chromosome.

infos.pos
Vector of integers specifying the physical position on a chromosome (in base pairs) of each SNP.
Typically <bigSNP>$map$physical.pos.

nploidy
Number of trials, parameter of the binomial distribution. Default is 2, which corresponds to diploidy, such as for the human genome.

ind.row
An optional vector of the row indices (individuals) that are used. If not specified, all rows are used.
Don't use negative indices.

ind.col
An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used.
Don't use negative indices.

ncores
Number of cores used. Default doesn't use parallelism. You may use nb_cores.

See also