bigsnpr: Predictive Analysis of Massive SNP Arrays

Easy-to-use, efficient, flexible and scalable tools for predictive analysis based on SNP arrays.

Arguments

G

A BM.code.descriptor (typically <bigSNP>$genotypes). You shouldn't have missing values in your data or SNPs with 0 MAF.

Gna

A BM.code.descriptor (typically <bigSNP>$genotypes). You can have missing values in your data.

x

A bigSNP.

infos.chr

Vector of integers specifying each SNP's chromosome. Typically <bigSNP>$map$chromosome.

infos.pos

Vector of integers specifying the physical position on a chromosome (in base pairs) of each SNP. Typically <bigSNP>$map$physical.pos.

nploidy

Number of trials, parameter of the binomial distribution. Default is 2, which corresponds to diploidy, such as for the human genome.

ind.row

An optional vector of the row indices (individuals) that are used. If not specified, all rows are used. Don't use negative indices.

ind.col

An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used. Don't use negative indices.

ncores

Number of cores used. Default doesn't use parallelism.

See also

Useful links:

https://privefl.github.io/bigsnpr

Report bugs at https://github.com/privefl/bigsnpr/issues