LD pruning and clumping

For a bigSNP:

snp_pruning: LD pruning. Similar to "--indep-pairwise (size+1) 1 thr.r2" in https://www.cog-genomics.org/plink/1.9/ld. (step is fixed to 1).

snp_clumping: LD clumping.

snp_indLRLDR: Get SNP indices of long-range LD regions for the human genome.

snp_clumping(G, infos.chr, ind.row = rows_along(G), S = NULL, size = 500,
  is.size.in.bp = FALSE, infos.pos = NULL, thr.r2 = 0.2, exclude = NULL,
  ncores = 1)

snp_pruning(G, infos.chr, ind.row = rows_along(G), size = 49,
  is.size.in.bp = FALSE, infos.pos = NULL, thr.r2 = 0.2, exclude = NULL,
  nploidy = getOption("bigsnpr.nploidy"), ncores = 1)

snp_indLRLDR(infos.chr, infos.pos, LD.regions = LD.wiki34)

Arguments

G

A BM.code.descriptor (typically <bigSNP>$genotypes). You shouldn't have missing values in your data or SNPs with 0 MAF.

infos.chr

Vector of integers specifying each SNP's chromosome. Typically <bigSNP>$map$chromosome.

ind.row

An optional vector of the row indices (individuals) that are used. If not specified, all rows are used. Don't use negative indices.

S

A vector of column statistics which express the importance of each SNP (the more important is the SNP, the greater should be the corresponding statistic). For example, if S follows the standard normal distribution, and "important" means significantly different from 0, you must use abs(S) instead. If not specified, the MAF is computed and used.

size

For one SNP, number of SNPs at its left and its right to be tested for being correlated with this particular SNP. This parameter should be adjusted with respect to the number of SNPs. The default are

49 for pruning (as in PLINK),

500 for clumping (I use this for a chip of 500K SNPs).

is.size.in.bp

Is the specified size is corresponding to kilo base-pairs. Default is FALSE (it is only SNP counts).

infos.pos

Vector of integers specifying the physical position on a chromosome (in base pairs) of each SNP. Typically <bigSNP>$map$physical.pos.

thr.r2

Threshold over the squared correlation between two SNPs. Default is 0.2.

exclude

Vector of SNP indices to exclude anyway. For example, can be used to exclude long-range LD regions (see Price2008). Another use can be for thresholding with respect to p-values associated with S.

ncores

Number of cores used. Default doesn't use parallelism.

nploidy

Number of trials, parameter of the binomial distribution. Default is 2, which corresponds to diploidy, such as for the human genome.

LD.regions

A data.frame with columns "Chr", "Start" and "Stop". Default use the table of 34 long-range LD regions that you can find https://goo.gl/0Ou7uI.

Value

snp_pruning & snp_pruning: SNP indices which are kept.

snp_indLRLDR: SNP indices to be used as (part of) the exclude parameter of snp_pruning or snp_clumping.

Details

I recommend to use clumping rather than pruning. See https://privefl.github.io/bigsnpr/articles/pruning-vs-clumping.html.

References

Price AL, Weale ME, Patterson N, et al. Long-Range LD Can Confound Genome Scans in Admixed Populations. Am J Hum Genet. 2008;83(1):132-135. http://dx.doi.org/10.1016/j.ajhg.2008.06.005

Examples

set.seed(1)

test <- snp_attachExtdata()
G <- test$genotypes
n <- nrow(G)
m <- ncol(G)

# pruning / clumping with MAF
ind.keep <- snp_pruning(G, infos.chr = test$map$chromosome, thr.r2 = 0.1)
# keep most of them -> not much LD in this simulated dataset
length(ind.keep) / m

#> [1] 0.9817261

ind.keep2 <- snp_clumping(G, infos.chr = test$map$chromosome, thr.r2 = 0.1)