MAX3 statistic

Counts the number of 0, 1, 2 by SNP and phenotype (cases/controls) and then compute the MAX3 statistic.

snp_MAX3(x, ind.train = seq(nrow(X)), val = c(0, 0.5, 1))

Arguments

x

A bigSNP.

ind.train

An optional vector of the row indices that are used, for the training part. If not specified, all data are used.

val

Computing ZCATT(x) for x in val. Default is c(0, 0.5, 1) and corresponds to the MAX3 statistic. Only c(0, 1) corresponds to MAX2. And only 0.5 corresponds to the Armatige trend test. seq(0, 1, length.out = L) corresponds to MAXL.

Value

A named list of S and pS for every column, which are MAX3 statistics and associated p-values. P-values are in fact the minimum of the 3 p-values of each test separately. One can use genomic control to rescale these p-values.

References

Zheng, G., Yang, Y., Zhu, X., & Elston, R. (2012). Robust Procedures. Analysis Of Genetic Association Studies, 151-206. http://dx.doi.org/10.1007/978-1-4614-2245-7_6.

Examples

set.seed(1)

# constructing a fake genotype big.matrix
a <- big.matrix(10, 15, type = "char", shared = FALSE)
a[] <- sample(c(0, 1, 2), 150, TRUE)

#> Assignment will down cast from double to char
#> Hint: To remove this warning type:  options(bigmemory.typecast.warning=FALSE)

#> Warning: 

print(a[,])

#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    0    0    2    1    2    1    2    1    1     0     1     2     2
#>  [2,]    1    0    0    1    1    2    0    2    2     0     1     2     1
#>  [3,]    1    2    1    1    2    1    1    1    1     1     0     1     1
#>  [4,]    2    1    0    0    1    0    0    1    0     2     2     1     0
#>  [5,]    0    2    0    2    1    0    1    1    2     2     1     0     2
#>  [6,]    2    1    1    2    2    0    0    2    0     2     0     0     1
#>  [7,]    2    2    0    2    0    0    1    2    2     1     0     2     1
#>  [8,]    1    2    1    0    1    1    2    1    0     1     1     0     0
#>  [9,]    1    1    2    2    2    1    0    2    0     2     2     1     0
#> [10,]    0    2    1    1    2    1    2    2    0     1     1     1     1
#>       [,14] [,15]
#>  [1,]     1     2
#>  [2,]     0     1
#>  [3,]     0     0
#>  [4,]     1     0
#>  [5,]     2     2
#>  [6,]     1     1
#>  [7,]     1     0
#>  [8,]     1     2
#>  [9,]     2     0
#> [10,]     1     2

# constructing a fake incomplete bigSNP with 10 individuals and 15 SNPs
# where the 5 first individuals are cases and the 5 last are controls.
fake <- list()
class(fake) <- "bigSNP"
fake$genotypes <- a
fake$fam$affection <- c(rep(1, 5), rep(-1, 5))

# Get MAX3 statistics
print(snp_MAX3(fake))

#> $S
#>  [1] 0.6666667 2.5000000 1.6666667 1.6666667 1.1111111 1.1111111 0.4761905
#>  [8] 3.6000000 3.6000000 2.5000000 0.4761905 0.6666667 2.5000000 2.5000000
#> [15] 0.0000000
#> 
#> $pS
#>  [1] 0.41421618 0.11384630 0.19670560 0.19670560 0.29184055 0.29184055
#>  [7] 0.49015296 0.05777957 0.05777957 0.11384630 0.49015296 0.41421618
#> [13] 0.11384630 0.11384630 1.00000000
#>