Attach a "bigSNP" from backing files

Load a bigSNP from backing files into R.

snp_attach(rdsfile)

Arguments

rdsfile

The path of the ".rds" which stores the bigSNP object.

Value

The bigSNP object.

Details

This is often just a call to readRDS. But it also checks if you have moved the two (".bk" and ".rds") backing files to another directory.

Examples

(bedfile <- system.file("extdata", "example.bed", package = "bigsnpr"))

#> [1] "/home/privef/Bureau/thesis/bigsnpr/inst/extdata/example.bed"

# Reading the bedfile and storing the data in directory "backingfiles"
rds <- snp_readBed(bedfile, backingfile = "test_doc")

#> Error: File 'backingfiles/test_doc.bk' already exists.

# Loading the data from backing files
test <- snp_attach(rds)

#> Error in path.expand(rdsfile): objet 'rds' introuvable

str(test)

#> Error in str(test): objet 'test' introuvable

dim(G <- test$genotypes)

#> Error in eval(expr, envir, enclos): objet 'test' introuvable

attach.BM(G)[1:8, 1:8]

#> Error in attach.BM(G): objet 'G' introuvable

# cleaning
bk <- sub("\\.rds$", ".bk", rds)

#> Error in sub("\\.rds$", ".bk", rds): objet 'rds' introuvable

file.remove(rds, bk)

#> Error in file.remove(rds, bk): objet 'rds' introuvable