Imputation

Imputation using Beagle version 4.

snp_beagleImpute(
  beagle.path,
  plink.path,
  bedfile.in,
  bedfile.out = NULL,
  memory.max = 3,
  ncores = 1,
  extra.options = "",
  plink.options = "",
  verbose = TRUE
)

Arguments

beagle.path
Path to the executable of Beagle v4+.

plink.path
Path to the executable of PLINK 1.9.

bedfile.in
Path to the input bedfile.

bedfile.out
Path to the output bedfile. Default is created by appending "_impute" to prefix.in (bedfile.in without extension).

memory.max
Max memory (in GB) to be used. It is internally rounded to be an integer. Default is 3.

ncores
Number of cores used. Default doesn't use parallelism. You may use nb_cores.

extra.options
Other options to be passed to Beagle as a string. More options can be found at Beagle's website.

plink.options
Other options to be passed to PLINK as a string. More options can be found at https://www.cog-genomics.org/plink2/filter.

verbose
Whether to show PLINK log? Default is TRUE.

Value

The path of the new bedfile.

Details

Downloads and more information can be found at the following websites

PLINK,

Beagle.

References

Browning, Brian L., and Sharon R. Browning. "Genotype imputation with millions of reference samples." The American Journal of Human Genetics 98.1 (2016): 116-126.

See also