Correlation

Get significant correlations between nearby SNPs.

snp_cor(G, ind.row = rows_along(G), ind.col = cols_along(G), size = 500,
  alpha = 0.05, fill.diag = TRUE)

Arguments

G
A BM.code.descriptor (typically <bigSNP>$genotypes). You shouldn't have missing values in your data or SNPs with 0 MAF.

ind.row
An optional vector of the row indices (individuals) that are used. If not specified, all rows are used. Don't use negative indices.

ind.col
An optional vector of the column indices (SNPs) that are used. If not specified, all columns are used. Don't use negative indices.

size
For one SNP, number of SNPs at its left and its right to be tested for being correlated with this particular SNP.

alpha
Type-I error for testing correlations.

fill.diag
Whether to fill the diagonal with 1s (the default) or to keep it as 0s.

Value

The correlation matrix. This is a sparse symmetric matrix.

Details

P-values are computed by a two-sided t-test.

Examples

test <- snp_attachExtdata()

corr <- snp_cor(test$genotypes)
corr[1:10, 1:10]

#> 10 x 10 sparse Matrix of class "dsCMatrix"
#>                                                                      
#>  [1,] 1.0000000 .  .           .          . . .  .          0.1113498
#>  [2,] .         1  .           .          . . .  .          .        
#>  [3,] .         .  1.00000000 -0.10260676 . . . -0.08716539 .        
#>  [4,] .         . -0.10260676  1.00000000 . . .  .          .        
#>  [5,] .         .  .           .          1 . .  .          .        
#>  [6,] .         .  .           .          . 1 .  .          .        
#>  [7,] .         .  .           .          . . 1  .          .        
#>  [8,] .         . -0.08716539  .          . . .  1.00000000 .        
#>  [9,] 0.1113498 .  .           .          . . .  .          1.0000000
#> [10,] .         .  .          -0.09303201 . . .  .          .        
#>                  
#>  [1,]  .         
#>  [2,]  .         
#>  [3,]  .         
#>  [4,] -0.09303201
#>  [5,]  .         
#>  [6,]  .         
#>  [7,]  .         
#>  [8,]  .         
#>  [9,]  .         
#> [10,]  1.00000000

length(corr@x) / length(corr)

#> [1] 0.009296616