Fake a "bigSNP"

Fake a "bigSNP"

snp_fake(n, m)

Arguments

n
Number of individuals.

m
Number of SNPs.

Value

A new temporary bigSNP object representing n individuals and m SNPs. The genotype Filebacked Big Matrix is initialized with missing values.

Examples

(test <- snp_fake(5, 12))

#> $genotypes
#> A Filebacked Big Matrix of type 'code 256' with 5 rows and 12 columns.
#> 
#> $fam
#>   family.ID sample.ID paternal.ID maternal.ID sex affection
#> 1         0     ind_1           0           0   0        -9
#> 2         0     ind_2           0           0   0        -9
#> 3         0     ind_3           0           0   0        -9
#> 4         0     ind_4           0           0   0        -9
#> 5         0     ind_5           0           0   0        -9
#> 
#> $map
#>    chromosome marker.ID genetic.dist physical.pos allele1 allele2
#> 1           1     snp_1            0            0       T       A
#> 2           1     snp_2            0            0       A       T
#> 3           1     snp_3            0            0       A       T
#> 4           1     snp_4            0            0       T       A
#> 5           1     snp_5            0            0       A       T
#> 6           1     snp_6            0            0       T       A
#> 7           1     snp_7            0            0       T       A
#> 8           1     snp_8            0            0       T       A
#> 9           1     snp_9            0            0       A       T
#> 10          1    snp_10            0            0       A       T
#> 11          1    snp_11            0            0       A       T
#> 12          1    snp_12            0            0       A       T
#> 
#> attr(,"class")
#> [1] "bigSNP"

# The genotype Filebackeg Big Matrix is initialized with missing values
G <- test$genotypes
G[]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#> [1,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [2,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [3,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [4,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA
#> [5,]   NA   NA   NA   NA   NA   NA   NA   NA   NA    NA    NA    NA

# Modify the genotype `big.matrix`
G[] <- sample(as.raw(0:3), size = length(G), replace = TRUE)
G[]

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12]
#> [1,]    0    0    2    1    0    2    2   NA    0     1     0     2
#> [2,]   NA    2    1   NA    1    2    2    0    2     1     0     2
#> [3,]    1    2    2    1    0   NA    2   NA   NA     2     0     1
#> [4,]   NA   NA   NA    1    0   NA   NA    2   NA    NA     1     1
#> [5,]    1    1    0    1    2    0    0    0    1     1    NA    NA