Get sample information

Get information of individuals by matching from an external file.

snp_getSampleInfos(x, df.or.files, col.family.ID = 1,
  col.sample.ID = 2, col.infos = -c(1, 2), pair.sep = "-_-", ...)

Arguments

x
A bigSNP.

df.or.files
Either

A data.frame,

A character vector of file names where to find at the information you want. You should have one column for family IDs and one for sample IDs.

col.family.ID
Index of the column containing the family IDs to match with those of the study. Default is the second one.

col.sample.ID
Index of the column containing the sample IDs to match with those of the study. Default is the first one.

col.infos
Indices of the column containing the information you want. Default is all but the first and the second columns.

pair.sep
Separator used for concatenation family and sample IDs in order to match easier. Default is "-_-".

...
Any additional parameter to pass to fread. Particularly, option header = FALSE is sometimes needed.

Value

The requested information as a data.frame.

See also

Examples

test <- snp_attachExtdata()
# Just after reading
rle(test$fam$family.ID)

#> Run Length Encoding
#>   lengths: int [1:19] 34 28 15 35 31 25 30 35 35 20 ...
#>   values : chr [1:19] "POP1" "POP2" "POP3" "POP4" "POP5" "POP6" "POP7" "POP8" "POP9" ...

# Get populations clusters from external files
files <- system.file("extdata", paste0("cluster", 1:3), package = "bigsnpr")
data.table::fread(files[1])

#>      POP1   IND0 POP-1-5
#>   1: POP1   IND1 POP-1-5
#>   2: POP1   IND2 POP-1-5
#>   3: POP1   IND3 POP-1-5
#>   4: POP1   IND4 POP-1-5
#>   5: POP1   IND5 POP-1-5
#>  ---                    
#> 138: POP5 IND138 POP-1-5
#> 139: POP5 IND139 POP-1-5
#> 140: POP5 IND140 POP-1-5
#> 141: POP5 IND141 POP-1-5
#> 142: POP5 IND142 POP-1-5

# need header option
data.table::fread(files[1], header = FALSE)

#>        V1     V2      V3
#>   1: POP1   IND0 POP-1-5
#>   2: POP1   IND1 POP-1-5
#>   3: POP1   IND2 POP-1-5
#>   4: POP1   IND3 POP-1-5
#>   5: POP1   IND4 POP-1-5
#>  ---                    
#> 139: POP5 IND138 POP-1-5
#> 140: POP5 IND139 POP-1-5
#> 141: POP5 IND140 POP-1-5
#> 142: POP5 IND141 POP-1-5
#> 143: POP5 IND142 POP-1-5

infos <- snp_getSampleInfos(test, files, header = FALSE)
rle(infos[[1]])

#> Run Length Encoding
#>   lengths: int [1:3] 143 167 207
#>   values : chr [1:3] "POP-1-5" "POP-6-11" "POP-12-19"