lassosum2

lassosum2

snp_lassosum2(
  corr,
  df_beta,
  delta = signif(seq_log(0.001, 3, 6), 1),
  nlambda = 20,
  lambda.min.ratio = 0.01,
  dfmax = 2e+05,
  maxiter = 500,
  tol = 1e-05,
  ncores = 1
)

Arguments

corr
Sparse correlation matrix as an SFBM. If corr is a dsCMatrix or a dgCMatrix, you can use as_SFBM(corr).

df_beta
A data frame with 3 columns:

$beta: effect size estimates

$beta_se: standard errors of effect size estimates

$n_eff: sample size when estimating beta (in the case of binary traits, this is 4 / (1 / n_control + 1 / n_case))

delta
Vector of shrinkage parameters to try (L2-regularization). Default is c(0.001, 0.005, 0.02, 0.1, 0.6, 3).

nlambda
Number of different lambdas to try. Default is 20.

lambda.min.ratio
Ratio between last and first lambdas to try. Default is 0.01.

dfmax
Maximum number of non-zero effects in the model. Default is 200e3.

maxiter
Maximum number of iterations before convergence. Default is 500.

tol
Tolerance parameter for assessing convergence. Default is 1e-5.

ncores
Number of cores used. Default doesn't use parallelism. You may use nb_cores.

Value

A matrix of effect sizes, one vector (column) for each row in attr(<res>, "grid_param"). Missing values are returned when strong divergence is detected.