Read BGEN files into a "bigSNP"

Function to read the UK Biobank BGEN files into a bigSNP.

snp_readBGEN(bgenfiles, backingfile, list_snp_id, ind_row = NULL,
  bgi_dir = dirname(bgenfiles), ncores = 1)

Arguments

bgenfiles
Character vector of paths to files with extension ".bgen". The corresponding ".bgen.bgi" index files must exist.

backingfile
The path (without extension) for the backing files for the cache of the bigSNP object.

list_snp_id
List (same length as the number of BGEN files) of character vector of SNP IDs to read. These should be in the form "<chr>_<pos>_<a1>_<a2>" (e.g. "1_88169_C_T" or "01_88169_C_T"). This function assumes that these IDs are uniquely identifying variants.

ind_row
An optional vector of the row indices (individuals) that are used. If not specified, all rows are used.
Don't use negative indices.

bgi_dir
Directory of index files. Default is the same as bgenfiles.

ncores
Number of cores used. Default doesn't use parallelism. You may use nb_cores().

Value

The path to the RDS file that stores the bigSNP object. Note that this function creates one other file which stores the values of the Filebacked Big Matrix.
You shouldn't read from BGEN files more than once. Instead, use snp_attach to load the "bigSNP" object in any R session from backing files.

Details

For more information on this format, please visit BGEN webpage.

This function is designed to read UK Biobank imputation files. This assumes that variants have been compressed with zlib, that there are only 2 possible alleles, and that each probability is stored on 8 bits.

Examples

# See e.g. https://github.com/privefl/UKBiobank/blob/master/10-get-dosages.R