Read PLINK files into a "bigSNP"

Function to read bed/bim/fam files into a bigSNP.

snp_readBed(bedfile, backingfile, backingpath = "backingfiles")

Arguments

bedfile
Path to file with extension ".bed" to read. You need the corresponding ".bim" and ".fam" in the same directory.

backingfile
The root name for the backing file(s) for the cache of the bigSNP object.

backingpath
The path to the directory containing the backing files. Default is "backingfiles".

Value

The path to the RDS file that stores the bigSNP object. Note that this function creates two other files which store the big.matrix and its descriptor. You shouldn't read from PLINK files more than once. Instead, use snp_attach to load the "bigSNP" object in any R session from backing files.

Details

For more information on these formats, please visit PLINK webpage. For other formats, please use PLINK to convert them in bedfiles, which require minimal space to store and are faster to read. For example, to convert from a VCF file, use the --vcf option.

Examples

(bedfile <- system.file("extdata", "example.bed", package = "bigsnpr"))

#> [1] "/home/privef/Bureau/bigsnpr/inst/extdata/example.bed"

# Reading the bedfile and storing the data in directory "backingfiles"
rds <- snp_readBed(bedfile, backingfile = "test_doc")

#> Error: File 'backingfiles/test_doc.bk' already exists.

# Loading the data from backing files
test <- snp_attach(rds)

#> Error in path.expand(rdsfile): objet 'rds' introuvable

str(test)

#> Error in str(test): objet 'test' introuvable

dim(G <- test$genotypes)

#> Error in eval(expr, envir, enclos): objet 'test' introuvable

attach.BM(G)[1:8, 1:8]

#> Error in attach.BM(G): objet 'G' introuvable

# cleaning
bk <- sub("\\.rds$", ".bk", rds)

#> Error in sub("\\.rds$", ".bk", rds): objet 'rds' introuvable

file.remove(rds, bk)

#> Error in file.remove(rds, bk): objet 'rds' introuvable