Save modifications

Save a bigSNP after having made some modifications to it. As bigSNP is an S3 class, you can add any slot you want to an object of this class, then use snp_saveModifs to save these modifications in the corresponding ".rds" backing file.

snp_saveModifs(x)

Arguments

x

A bigSNP.

Value

The (saved) bigSNP.

Examples

set.seed(1)

#reading example
test <- snp_attachExtdata()

# I can add whatever I want to an S3 class
test$map$`p-values` <- runif(nrow(test$map))
str(test$map)

#> 'data.frame':	4542 obs. of  7 variables:
#>  $ chromosome  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ marker.ID   : chr  "SNP0" "SNP1" "SNP2" "SNP3" ...
#>  $ genetic.dist: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ physical.pos: int  0 1000 2000 3000 4000 5000 6000 7000 8000 9000 ...
#>  $ allele1     : chr  "A" "T" "T" "T" ...
#>  $ allele2     : chr  "T" "A" "A" "A" ...
#>  $ p-values    : num  0.2672 0.3861 0.0134 0.3824 0.8697 ...

# reading again
test2 <- snp_attach(rdsfile = test$savedIn)
str(test2$map) # new slot wasn't saved

#> 'data.frame':	4542 obs. of  6 variables:
#>  $ chromosome  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ marker.ID   : chr  "SNP0" "SNP1" "SNP2" "SNP3" ...
#>  $ genetic.dist: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ physical.pos: int  0 1000 2000 3000 4000 5000 6000 7000 8000 9000 ...
#>  $ allele1     : chr  "A" "T" "T" "T" ...
#>  $ allele2     : chr  "T" "A" "A" "A" ...

# save it
test2$map$`p-values` <- runif(nrow(test2$map))
test2 <- snp_saveModifs(test2)

# reading again
test3 <- snp_attach(rdsfile = test2$savedIn)
str(test3$map) # it is saved now

#> 'data.frame':	4542 obs. of  7 variables:
#>  $ chromosome  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ marker.ID   : chr  "SNP0" "SNP1" "SNP2" "SNP3" ...
#>  $ genetic.dist: int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ physical.pos: int  0 1000 2000 3000 4000 5000 6000 7000 8000 9000 ...
#>  $ allele1     : chr  "A" "T" "T" "T" ...
#>  $ allele2     : chr  "T" "A" "A" "A" ...
#>  $ p-values    : num  0.643 0.849 0.455 0.872 0.15 ...

# the complicated code of this function
snp_saveModifs

#> function(x) {
#>   saveRDS(x, x$savedIn)
#> 
#>   x
#> }
#> <environment: namespace:bigsnpr>