Split-parApply-Combine

A Split-Apply-Combine strategy to parallelize the evaluation of a function on each SNP, independently.

snp_split(infos.chr, FUN, combine, ncores = 1, ...)

Arguments

infos.chr

Vector of integers specifying the chromosome of each SNP. Typically the slot chromosome of the slot map of a bigSNP.

FUN

The function to be applied. It must take a BM.code as first argument and ind.chr, an another argument to provide subsetting over SNPs. You can access the number of the chromosome by using attr(ind.chr, "chr").

combine

function that is used by foreach to process the tasks results as they generated. This can be specified as either a function or a non-empty character string naming the function. Specifying 'c' is useful for concatenating the results into a vector, for example. The values 'cbind' and 'rbind' can combine vectors into a matrix. The values '+' and '*' can be used to process numeric data. By default, the results are returned in a list.

ncores

Number of cores used. Default doesn't use parallelism.

...

Extra arguments to be passed to FUN.

Value

The result of foreach.

Details

This function splits indices for each chromosome, then apply a given function to each part (chromosome) and finally combine the results.