Subset

A method to get a subset (copy) of a bigSNP. This new bigSNP will also be backed by files (in the same directory).

# S3 method for bigSNP
subset(x, ind.row = rows_along(X), ind.col = cols_along(X),
  ...)

Arguments

x
A bigSNP.

ind.row
Indices of the rows (individuals) to keep. Negative indices can be used to exclude row indices. Default: keep them all.

ind.col
Indices of the columns (SNPs) to keep. Negative indices can be used to exclude column indices. Default: keep them all.

...
Not used.

Value

The path to the RDS file that stores the bigSNP object. Note that this function creates two files whose names have been automatically chosen by appending "_sub" and a number to the prefix of the input bigSNP backing files.

See also

bigSNP

Examples

str(test <- snp_attachExtdata())

#> List of 4
#>  $ genotypes:Formal class 'BM.code.descriptor' [package "bigstatsr"] with 2 slots
#>   .. ..@ code       : num [1:256] 0 1 2 NA NA NA NA NA NA NA ...
#>   .. ..@ description:List of 13
#>   .. .. ..$ sharedType: chr "FileBacked"
#>   .. .. ..$ filename  : chr "file3a3a1a2f642d.bk"
#>   .. .. ..$ dirname   : chr "/tmp/RtmpbxbvnH/"
#>   .. .. ..$ totalRows : num 517
#>   .. .. ..$ totalCols : num 4542
#>   .. .. ..$ rowOffset : num 0
#>   .. .. ..$ colOffset : num 0
#>   .. .. ..$ nrow      : num 517
#>   .. .. ..$ ncol      : num 4542
#>   .. .. ..$ rowNames  : NULL
#>   .. .. ..$ colNames  : NULL
#>   .. .. ..$ type      : chr "raw"
#>   .. .. ..$ separated : logi FALSE
#>  $ fam      :'data.frame':	517 obs. of  6 variables:
#>   ..$ family.ID  : chr [1:517] "POP1" "POP1" "POP1" "POP1" ...
#>   ..$ sample.ID  : chr [1:517] "IND0" "IND1" "IND2" "IND3" ...
#>   ..$ paternal.ID: int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ maternal.ID: int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ sex        : int [1:517] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ affection  : int [1:517] 1 1 2 1 1 1 1 1 1 1 ...
#>  $ map      :'data.frame':	4542 obs. of  6 variables:
#>   ..$ chromosome  : int [1:4542] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ marker.ID   : chr [1:4542] "SNP0" "SNP1" "SNP2" "SNP3" ...
#>   ..$ genetic.dist: int [1:4542] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ physical.pos: int [1:4542] 112 1098 2089 3107 4091 5091 6107 7103 8090 9074 ...
#>   ..$ allele1     : chr [1:4542] "A" "T" "T" "T" ...
#>   ..$ allele2     : chr [1:4542] "T" "A" "A" "A" ...
#>  $ savedIn  : chr "/tmp/RtmpbxbvnH/file3a3a1a2f642d.rds"
#>  - attr(*, "class")= chr "bigSNP"

# keep only first 50 samples and SNPs
rdsfile <- subset(test, ind.row = 1:50, ind.col = 1:50)
str(snp_attach(rdsfile))

#> List of 4
#>  $ genotypes:Formal class 'BM.code.descriptor' [package "bigstatsr"] with 2 slots
#>   .. ..@ code       : num [1:256] 0 1 2 NA NA NA NA NA NA NA ...
#>   .. ..@ description:List of 13
#>   .. .. ..$ sharedType: chr "FileBacked"
#>   .. .. ..$ filename  : chr "file3a3a1a2f642d_sub1.bk"
#>   .. .. ..$ dirname   : chr "/tmp/RtmpbxbvnH/"
#>   .. .. ..$ totalRows : num 50
#>   .. .. ..$ totalCols : num 50
#>   .. .. ..$ rowOffset : num 0
#>   .. .. ..$ colOffset : num 0
#>   .. .. ..$ nrow      : num 50
#>   .. .. ..$ ncol      : num 50
#>   .. .. ..$ rowNames  : NULL
#>   .. .. ..$ colNames  : NULL
#>   .. .. ..$ type      : chr "raw"
#>   .. .. ..$ separated : logi FALSE
#>  $ fam      :'data.frame':	50 obs. of  6 variables:
#>   ..$ family.ID  : chr [1:50] "POP1" "POP1" "POP1" "POP1" ...
#>   ..$ sample.ID  : chr [1:50] "IND0" "IND1" "IND2" "IND3" ...
#>   ..$ paternal.ID: int [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ maternal.ID: int [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ sex        : int [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ affection  : int [1:50] 1 1 2 1 1 1 1 1 1 1 ...
#>  $ map      :'data.frame':	50 obs. of  6 variables:
#>   ..$ chromosome  : int [1:50] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ marker.ID   : chr [1:50] "SNP0" "SNP1" "SNP2" "SNP3" ...
#>   ..$ genetic.dist: int [1:50] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ physical.pos: int [1:50] 112 1098 2089 3107 4091 5091 6107 7103 8090 9074 ...
#>   ..$ allele1     : chr [1:50] "A" "T" "T" "T" ...
#>   ..$ allele2     : chr [1:50] "T" "A" "A" "A" ...
#>  $ savedIn  : chr "/tmp/RtmpbxbvnH/file3a3a1a2f642d_sub1.rds"
#>  - attr(*, "class")= chr "bigSNP"

# remove only first 50 samples and SNPs
rdsfile2 <- subset(test, ind.row = -(1:50), ind.col = -(1:50))
str(snp_attach(rdsfile2))

#> List of 4
#>  $ genotypes:Formal class 'BM.code.descriptor' [package "bigstatsr"] with 2 slots
#>   .. ..@ code       : num [1:256] 0 1 2 NA NA NA NA NA NA NA ...
#>   .. ..@ description:List of 13
#>   .. .. ..$ sharedType: chr "FileBacked"
#>   .. .. ..$ filename  : chr "file3a3a1a2f642d_sub2.bk"
#>   .. .. ..$ dirname   : chr "/tmp/RtmpbxbvnH/"
#>   .. .. ..$ totalRows : num 467
#>   .. .. ..$ totalCols : num 4492
#>   .. .. ..$ rowOffset : num 0
#>   .. .. ..$ colOffset : num 0
#>   .. .. ..$ nrow      : num 467
#>   .. .. ..$ ncol      : num 4492
#>   .. .. ..$ rowNames  : NULL
#>   .. .. ..$ colNames  : NULL
#>   .. .. ..$ type      : chr "raw"
#>   .. .. ..$ separated : logi FALSE
#>  $ fam      :'data.frame':	467 obs. of  6 variables:
#>   ..$ family.ID  : chr [1:467] "POP2" "POP2" "POP2" "POP2" ...
#>   ..$ sample.ID  : chr [1:467] "IND50" "IND51" "IND52" "IND53" ...
#>   ..$ paternal.ID: int [1:467] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ maternal.ID: int [1:467] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ sex        : int [1:467] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ affection  : int [1:467] 1 1 2 1 1 1 1 1 1 2 ...
#>  $ map      :'data.frame':	4492 obs. of  6 variables:
#>   ..$ chromosome  : int [1:4492] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ marker.ID   : chr [1:4492] "SNP50" "SNP51" "SNP52" "SNP53" ...
#>   ..$ genetic.dist: int [1:4492] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ physical.pos: int [1:4492] 50100 51109 52101 53106 54101 55097 56103 57110 58101 59088 ...
#>   ..$ allele1     : chr [1:4492] "A" "A" "A" "T" ...
#>   ..$ allele2     : chr [1:4492] "T" "T" "T" "A" ...
#>  $ savedIn  : chr "/tmp/RtmpbxbvnH/file3a3a1a2f642d_sub2.rds"
#>  - attr(*, "class")= chr "bigSNP"